- David H. Bronke merged changes from revisions 24811 through 24913 of 	  trunk into the streamloader branch.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/soc/streamloader@24914 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/imesh/emit.h
@@ -297,7 +297,7 @@ struct iEmitState : public virtual iBase
   /// Set the number of particles to use.
   virtual void SetParticleCount (int num) = 0;
   /// Get the number of particles used.
-  virtual int GetParticleCount () const = 0;
+  virtual size_t GetParticleCount () const = 0;
   /// Enable or disable lighting.
   virtual void SetLighting (bool l) = 0;
   /// See if lighting is enabled.
