+++ include/imesh/fire.h
@@ -42,7 +42,7 @@ struct iFireState : public virtual iBase
   /// Set the number of particles to use.
   virtual void SetParticleCount (int num) = 0;
   /// Get the number of particles.
-  virtual int GetParticleCount () const = 0;
+  virtual size_t GetParticleCount () const = 0;
   /// Set the size of the fire drops.
   virtual void SetDropSize (float dropwidth, float dropheight) = 0;
   /// Get the size of the fire drops.
